backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.0
openai==1.3.0
requests==2.31.0
pydub==0.25.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
